requirements.txt
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.29
django-ckeditor==5.4.0
django-debug-toolbar==2.1
django-js-asset==0.1.1
django-suit==0.2.26
idna==2.5
importlib-metadata==1.4.0
kombu==4.2.1
more-itertools==8.1.0
mysqlclient==1.3.9
Pillow==6.2.1
python-decouple==3.3
python-http-client==2.2.1
pytz==2018.3
requests==2.22.0
sendgrid==4.2.1
sqlparse==0.3.0
urllib3==1.25.7
vine==1.1.4
zipp==2.0.1
